moved navbar
